+++ pkg/reportingplugins/mercury/v0/observation.go
@@ -73,7 +73,7 @@ func (pao ParsedAttributedObservation) GetCurrentBlockValid() bool {
 	return pao.CurrentBlockValid
 }
 
-func (pao ParsedAttributedObservation) GetMaxFinalizedTimestamp() int64 {
+func (pao ParsedAttributedObservation) GetMaxFinalizedTimestamp() uint32 {
 	panic("current observation doesn't contain the field")
 }
